WIP: Rethink AutoPlotter as an abc.
