+++ bluesky_widgets/models/auto_plot_builders/__init__.py
@@ -1,0 +1,3 @@
+from .base import *
+from .lines import *
+from .images import *
